+++ code/src/sixsq/nuvla/auth/password.clj
@@ -2,6 +2,8 @@
   (:refer-clojure :exclude [update])
   (:require
     [buddy.hashers :as hashers]
+    [sixsq.nuvla.auth.acl-resource :as a]
+    [sixsq.nuvla.auth.utils :as auth]
     [sixsq.nuvla.db.filter.parser :as parser]
     [sixsq.nuvla.db.impl :as db]
     [sixsq.nuvla.server.resources.common.crud :as crud]
@@ -74,3 +76,28 @@
                               :hash)]
     (when (valid-password? password password-hash)
       user)))
+
+
+(defn check-email-validated
+  [{:keys [validated] :as email}]
+  (when validated
+    email))
+
+
+(defn user-id->email
+  [user-id]
+  (some-> user-id
+          crud/retrieve-by-id-as-admin
+          :email
+          crud/retrieve-by-id-as-admin
+          check-email-validated
+          :address))
+
+
+(defn invited-by
+  [request]
+  (let [user-id (auth/current-user-id request)]
+    (or (user-id->email user-id)
+        (if (a/is-admin? (auth/current-authentication request))
+          (some-> user-id user-id->user :name)
+          user-id))))